feat: add anti-aliasing with default 8 rays per pixel
add anti-aliasing in rendering
add toggle (G) to enable/disable anti-aliasing (disabled by default as AA renders are time consuming)
renders without AA when camera or object is moved using the mouse
add preprocessor `RAYS_PER_PIXEL` for the number of rays per pixel used for AA rendering

+++ src/camera/camera.c
@@ -6,7 +6,7 @@
 /*   By: fgargot <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/22 19:15:40 by fgargot           #+#    #+#             */
-/*   Updated: 2026/07/20 23:54:49 by fgargot          ###   ########.fr       */
+/*   Updated: 2026/08/01 17:25:47 by fgargot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ static t_vec3	get_ray_direction(t_object *cam, double u, double v)
 	return (dir);
 }
 
-t_ray	camera_ray(t_object *cam, int x, int y)
+t_ray	camera_ray(t_object *cam, double x, double y)
 {
 	double	u;
 	double	v;
